chore: updated changes to class
