+++ packages/starknet/lib/src/core/types/felt.dart
@@ -2,9 +2,8 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import '../convert.dart';
-import '../interfaces/starknet_serializable.dart';
 
-class Felt extends StarknetSerializable {
+class Felt implements IToCalldata {
   /// Spec: https://docs.starknet.io/docs/Hashing/hash-functions/#domain-and-range
   static final prime =
       BigInt.two.pow(251) + BigInt.from(17) * BigInt.two.pow(192) + BigInt.one;
@@ -155,5 +154,35 @@ extension Starknet on BigInt {
     }
 
     return data;
+  }
+}
+
+/// Converts an object to a list of `Felt`s suitable for Starknet calldata.
+///
+/// Implement this interface to allow your object to be converted to calldata.
+abstract interface class IToCalldata {
+  List<Felt> toCalldata();
+}
+
+extension SerializableListToCalldata<T extends IToCalldata> on List<T> {
+  List<Felt> toCalldata() {
+    if (isEmpty) return [Felt.zero];
+
+    return [
+      Felt.fromInt(length),
+      ...expand((item) => item.toCalldata()),
+    ];
+  }
+}
+
+extension NestedSerializableListToCalldata<T extends IToCalldata>
+    on List<List<T>> {
+  List<Felt> toCalldata() {
+    if (isEmpty) return [Felt.zero];
+
+    return [
+      Felt.fromInt(length),
+      ...expand((innerList) => innerList.toCalldata()),
+    ];
   }
 }
