+++ packages/starknet/lib/src/core/types/uint256.dart
@@ -1,14 +1,12 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-
-import '../interfaces/starknet_serializable.dart';
 import 'felt.dart';
 
 @immutable
-class Uint256 extends StarknetSerializable {
+class Uint256 implements IToCalldata {
   final Felt low; // low 128 bits
   final Felt high; // high 128 bits
 
-  Uint256({required this.low, required this.high});
+  const Uint256({required this.low, required this.high});
 
   factory Uint256.fromInt(int number) =>
       Uint256.fromBigInt(BigInt.from(number));
